projects/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Caching
Flask-Limiter
transformers
torch
pinecone-client
